Pay invoice from qr code  (#6)
