--- delete.go
@@ -1,7 +1,0 @@
-package main
-
-// func deleteMessage(update tgbotapi.Update, bot TipBot) error {
-// 	TippedMessage(update.Message, time.Millisecond*time.Duration(0)).Dispose(bot.telegram)
-// 	// m.Dispose(bot.telegram)
-// 	return nil
-// }
