+++ go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/imroc/req v0.3.0
 	github.com/jinzhu/configor v1.2.1
 	github.com/liyue201/goqr v0.0.0-20200803022322-df443203d4ea
+	github.com/makiuchi-d/gozxing v0.0.2
 	github.com/sirupsen/logrus v1.2.0
 	github.com/skip2/go-qrcode v0.0.0-20200617195104-da1b6568686e
 	gopkg.in/tucnak/telebot.v2 v2.3.5
