+++ internal/lnbits/lnbits.go
@@ -2,19 +2,9 @@ package lnbits
 
 import (
 	"github.com/imroc/req"
-	tb "gopkg.in/tucnak/telebot.v2"
 )
 
-type Client struct {
-	header     req.Header
-	url        string
-	AdminKey   string
-	InvoiceKey string
-}
-
-// NewClient is the first function you must call. Pass your main API key here.
-// It will return a client you can later use to access wallets and transactions.
-// You can find it at https://lnpay.co/developers/dashboard
+// NewClient returns a new lnbits api client. Pass your API key and url here.
 func NewClient(key, url string) *Client {
 	return &Client{
 		url: url,
@@ -26,30 +16,7 @@ func NewClient(key, url string) *Client {
 	}
 }
 
-type User struct {
-	ID          string       `json:"id"`
-	Name        string       `json:"name" gorm:"primaryKey"`
-	Initialized bool         `json:"initialized"`
-	Telegram    *tb.User     `gorm:"embedded;embeddedPrefix:telegram_"`
-	Wallet      *Wallet      `gorm:"embedded;embeddedPrefix:wallet_"`
-	StateKey    UserStateKey `json:"stateKey"`
-	StateData   string       `json:"stateData"`
-}
-type UserStateKey int
-
-const (
-	UserStateConfirmPayment = iota + 1
-	UserStateConfirmSend
-)
-
-func (u *User) ResetState() {
-	u.StateData = ""
-	u.StateKey = 0
-}
-
-// CreateWallet creates a new wallet with a given descriptive label.
-// It will return the wallet object which you can use to create invoices and payments.
-// https://docs.lnpay.co/wallet/create-wallet
+// GetUser returns user information
 func (c *Client) GetUser(userId string) (user User, err error) {
 	resp, err := req.Post(c.url+"/usermanager/api/v1/users/"+userId, c.header, nil)
 	if err != nil {
@@ -67,9 +34,7 @@ func (c *Client) GetUser(userId string) (user User, err error) {
 	return
 }
 
-// CreateWallet creates a new wallet with a given descriptive label.
-// It will return the wallet object which you can use to create invoices and payments.
-// https://docs.lnpay.co/wallet/create-wallet
+// CreateUserWithInitialWallet creates new user with initial wallet
 func (c *Client) CreateUserWithInitialWallet(userName, walletName, adminId string, email string) (wal User, err error) {
 	resp, err := req.Post(c.url+"/usermanager/api/v1/users", c.header, req.BodyJSON(struct {
 		WalletName string `json:"wallet_name"`
@@ -91,9 +56,7 @@ func (c *Client) CreateUserWithInitialWallet(userName, walletName, adminId strin
 	return
 }
 
-// CreateWallet creates a new wallet with a given descriptive label.
-// It will return the wallet object which you can use to create invoices and payments.
-// https://docs.lnpay.co/wallet/create-wallet
+// CreateWallet creates a new wallet.
 func (c *Client) CreateWallet(userId, walletName, adminId string) (wal Wallet, err error) {
 	resp, err := req.Post(c.url+"/usermanager/api/v1/wallets", c.header, req.BodyJSON(struct {
 		UserId     string `json:"user_id"`
@@ -115,15 +78,7 @@ func (c *Client) CreateWallet(userId, walletName, adminId string) (wal Wallet, e
 	return
 }
 
-type InvoiceParams struct {
-	Out     bool   `json:"out"`
-	Amount  int64  `json:"amount"`
-	Memo    string `json:"memo"` // the invoice description.
-	Webhook string `json:"webhook,omitempty"`
-}
-
 // Invoice creates an invoice associated with this wallet.
-// https://docs.lnpay.co/wallet/generate-invoice
 func (c Client) Invoice(params InvoiceParams, w Wallet) (lntx BitInvoice, err error) {
 	c.header["X-Api-Key"] = w.Adminkey
 	resp, err := req.Post(c.url+"/api/v1/payments", w.header, req.BodyJSON(&params))
@@ -142,18 +97,7 @@ func (c Client) Invoice(params InvoiceParams, w Wallet) (lntx BitInvoice, err er
 	return
 }
 
-type PaymentParams struct {
-	Out    bool   `json:"out"`
-	Bolt11 string `json:"bolt11"`
-}
-type PayParams struct {
-	// the BOLT11 payment request you want to pay.
-	PaymentRequest string `json:"payment_request"`
-
-	// custom data you may want to associate with this invoice. optional.
-	PassThru map[string]interface{} `json:"passThru"`
-}
-
+// Info returns wallet information
 func (c Client) Info(w Wallet) (wtx Wallet, err error) {
 	c.header["X-Api-Key"] = w.Adminkey
 	resp, err := req.Get(w.url+"/api/v1/wallet", w.header, nil)
@@ -171,6 +115,8 @@ func (c Client) Info(w Wallet) (wtx Wallet, err error) {
 	err = resp.ToJSON(&wtx)
 	return
 }
+
+// Wallets returns all wallets belonging to an user
 func (c Client) Wallets(w User) (wtx []Wallet, err error) {
 	resp, err := req.Get(c.url+"/usermanager/api/v1/wallets/"+w.ID, c.header, nil)
 	if err != nil {
@@ -189,7 +135,6 @@ func (c Client) Wallets(w User) (wtx []Wallet, err error) {
 }
 
 // Pay pays a given invoice with funds from the wallet.
-// https://docs.lnpay.co/wallet/pay-invoice
 func (c Client) Pay(params PaymentParams, w Wallet) (wtx BitInvoice, err error) {
 	c.header["X-Api-Key"] = w.Adminkey
 	resp, err := req.Post(c.url+"/api/v1/payments", w.header, req.BodyJSON(&params))
@@ -206,10 +151,4 @@ func (c Client) Pay(params PaymentParams, w Wallet) (wtx BitInvoice, err error) 
 
 	err = resp.ToJSON(&wtx)
 	return
-}
-
-type TransferParams struct {
-	Memo         string `json:"memo"`           // the transfer description.
-	NumSatoshis  int64  `json:"num_satoshis"`   // the transfer amount.
-	DestWalletId string `json:"dest_wallet_id"` // the key or id of the destination
 }
